✨🔧 Add Rust installation option
This commit adds the ability to install Rust as part of the installation process. Users will be prompted to choose whether they want to use Rust, and if they select 'y', the script will execute the necessary commands to install Rust using Homebrew. The script also ensures that Homebrew is installed and up-to-date before proceeding with the installation.

The new file `rust/install.sh` contains the code for installing Rust, including updating Homebrew, installing `rustup-init`, setting the default toolchain to nightly, and cleaning up outdated versions.