✨🔧 Refactor code structure and remove unnecessary files
This commit refactors the code structure by removing unused files. The `.gitmodules` file was deleted, as it is no longer needed. Additionally, the `nvim/.config/nvim` directory was removed as it contained a subproject that is no longer relevant.

The `_install/zsh.sh` script was modified to use the `stow` command to install zsh instead of directly referencing the dotfiles path. This change improves modularity and simplifies installation.

Lastly, a tool for cleaning up macOS, called `cleanup`, was added by downloading it from a GitHub repository.

These changes enhance code organization and improve installation processes.