✨🔧 Add git installation script and update git configuration
This commit adds a new bash script to install git. The script prints messages indicating the progress of the installation and uses `stow` to configure git. Additionally, the `.gitattributes` and `.gitignore` files are moved to the `.config/git/` directory.

The changes made to the `.gitconfig` file include:
- Indentation fix for `gpgsign`
- Added user information (name, email, signing key)
- Added GitHub credentials (username, token)
- Updated GPG program path
- Set default branch to `main`
- Disabled forced annotation signing for tags

The overall installation process is improved by sourcing the new git installation script in `install.sh`.