✨🔧 Add GOPATH and GOROOT environment variables
This commit adds the necessary environment variables for setting up the Go programming language. The GOPATH is set to "$HOME/go" and the GOROOT is set to "$HOME/.go". Additionally, the PATH variable is updated to include "$GOPATH/bin". Two aliases, "g-install" and "ggovm", are also added. For more information, refer to https://github.com/stefanmaric/g.