✨🔧 Add options to open files with different applications
This commit adds a feature that allows users to select how they want to open a file. The user is presented with a list of options, including opening the file with the default application, less, vim, cat, web browser (Google Chrome), WebStorm, IntelliJ IDEA, or CLion. The selected option will be used to open the chosen file.