✨🔧 Update plugin dependencies and configurations
This commit updates the dependencies and configurations for several plugins in the Neovim configuration. Notable changes include:

- Updated `conform.nvim` to commit `46c107ad0e7d83b5a7091112ec2994c847577d32`
- Updated `mason-lspconfig.nvim` to commit `25c11854aa25558ee6c03432edfa0df0217324be`
- Updated `nvim-lspconfig` to commit `d88ae6623fef09251e3aa20001bb761686eae730`
- Updated `nvim-tree.lua` to commit `d41b4ca013ed89e41b9c0ecbdae5f1633e42f7fa`
- Updated `nvim-treesitter` to commit `b6a6d8997c46dc15682020ce4fddc5a89ee1ac0d`
- Updated `nvim-web-devicons` to commit `9793801f974bba70e4ac5d7eae6c4f5659993d8e`

Additionally, the configuration for the plugin "rcarriga/nvim-notify" has been updated with a custom background color.

These changes ensure that the Neovim configuration is up-to-date with the latest versions of the plugins and their respective configurations.