✨🔧 Add functions.zsh and mkdock.zsh
This commit adds a new file `functions.zsh` to the zsh settings directory, which contains a function called `_mkdock`. The `_mkdock` function is responsible for customizing the dock by adding specific applications and folders. It also creates aliases `dock` and `mkdock` that can be used to execute the `_mkdock` function.