✨🔧 Add new aliases for terminal and NeoVim
This commit adds new aliases to the `aliases.zsh` file. The `cl` alias is added to clear the terminal in a short way, while the `vim` and `vi` aliases are added to use NeoVim as an alternative to Vim. These additions enhance productivity and provide more flexibility in the terminal environment.